requirements.txt
python-telegram-bot
langchain
chromadb
langchain_groq
langchain_huggingface
PyPDF2
sentence-transformers
python-dotenv
langchain_chroma

